pyproject.toml
[project]
name = "orders"
version = "0.1.0"
description = "Orders management system"
authors = [
    {name = "Fernando Pioli",email = "fernandohrp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]
[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
django-extensions = "^4.1"
django-debug-toolbar = "^5.2.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pytest-django = "^4.5.0"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "web.config.settings"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = ["--cov=src", "--cov-config=.coveragerc", "--verbose"]